+++ internal/capability/capability.go
@@ -29,7 +29,7 @@ func All() []Capability {
 // If the construct does not require a capability, it returns None.
 func ForConstruct(name string) Capability {
 	switch name {
-	case "db_connect", "sql_query", "store_open", "store_put", "store_get", "store_delete":
+	case "db_connect", "sql_query", "store_open", "store_put", "store_get", "store_delete", "store_keys":
 		return Database
 	case "fetch", "res", "res_json", "res_header", "http_res_header", "req_method", "http_req_method", "http_server_start", "http_route", "http_server_serve", "llm_generate", "achieve":
 		return Network